refactor: 서비스 분리에 따른 변경사항 반영
- 서비스 코드 수정
- DTO 위치 변경 및 커스텀 예외 객체 생성
- Controller 경로 수정
