+++ src/business/member/model/dto/member-login-request.dto.ts
@@ -10,4 +10,13 @@ export class MemberLoginRequestDto {
   @IsNotEmpty()
   @ApiProperty({ description: '비밀번호', example: '1234' })
   public readonly password: string;
+
+  constructor(email: string, password: string) {
+    this.email = email;
+    this.password = password;
+  }
+
+  public static of(email: string, password: string): MemberLoginRequestDto {
+    return new MemberLoginRequestDto(email, password);
+  }
 }
